requirements.txt
nest_asyncio
fastapi
uvicorn
python-dotenv
sentence-transformers
pinecone
pinecone-text
